fix: 라우팅 수정 #78
